+++ src/game_input.h
@@ -52,14 +52,14 @@ typedef enum
 } mouseaxis_e;
 
 extern uint32_t gamecontrols[MAX_PLAYERS][NUMCONTROLS];
-extern int16_t  gameaxis[MAX_PLAYERS][NUMAXIS];
+extern int32_t  gameaxis[MAX_PLAYERS][NUMAXIS];
 extern uint32_t gamecontrolbinds[MAX_PLAYERS][NUMCONTROLS][2];
 extern uint32_t mousecontrols[MAX_PLAYERS][NUMMOUSECONTROLS];
-extern int16_t  mouseaxis[MAX_PLAYERS][NUMMOUSEAXIS];
+extern int32_t  mouseaxis[MAX_PLAYERS][NUMMOUSEAXIS];
 
 void G_DefaultControls(void);
 bool G_ControlDown(uint8_t player, uint8_t control, bool pressed);
-int16_t G_PlayerAxis(uint8_t player, uint8_t axis);
-int16_t G_MouseAxis(uint8_t player, uint8_t axis);
+int32_t G_PlayerAxis(uint8_t player, uint8_t axis);
+int32_t G_MouseAxis(uint8_t player, uint8_t axis);
 
 #endif
